make application null-safe

+++ frontend/src/main/java/org/cc/torganizer/frontend/players/actions/DeletePlayer.java
@@ -1,6 +1,8 @@
 package org.cc.torganizer.frontend.players.actions;
 
 import javax.enterprise.context.RequestScoped;
+import javax.faces.application.FacesMessage;
+import javax.faces.context.FacesContext;
 import javax.inject.Inject;
 import javax.inject.Named;
 import org.apache.logging.log4j.Logger;
@@ -21,6 +23,12 @@ public class DeletePlayer extends PlayersAction {
     // delete from tournament and from players/persons-tables
     Long tId = appState.getCurrent().getId();
     Long pId = state.getCurrent().getId();
+
+    if (tId == null || pId == null) {
+      FacesMessage message = new FacesMessage("Fehbler beim löschen eines Players");
+      FacesContext.getCurrentInstance().addMessage(null, message);
+      return;
+    }
 
     tournamentsRepository.removePlayer(tId, pId);
     playersRepository.delete(pId);
